src/main/java/day_247/NumberOfWaysOfCuttingAPizza.java
package day_247;

/**
 * <a href = "https://leetcode.com/problems/number-of-ways-of-cutting-a-pizza/">
 * 1444. Number of Ways of Cutting a Pizza </a>
 * <p>
 * Given a rectangular pizza represented as a rows x cols matrix containing the following characters: 'A' (an apple) and '.' (empty cell) and given the integer k. You have to cut the pizza into k pieces using k-1 cuts.
 * <p>
 * For each cut you choose the direction: vertical or horizontal, then you choose a cut position at the cell boundary and cut the pizza into two pieces. If you cut the pizza vertically, give the left part of the pizza to a person. If you cut the pizza horizontally, give the upper part of the pizza to a person. Give the last piece of pizza to the last person.
 * <p>
 * Return the number of ways of cutting the pizza such that each piece contains at least one apple. Since the answer can be a huge number, return this modulo 10^9 + 7.
 */

public class NumberOfWaysOfCuttingAPizza {
    public int ways(String[] pizza, int k) {
        int m = pizza.length;
        int n = pizza[0].length();
        Integer[][][] dp = new Integer[k][m][n];
        int[][] preSum = new int[m + 1][n + 1];

        for (int r = m - 1; r >= 0; r--)
            for (int c = n - 1; c >= 0; c--)
                preSum[r][c] = preSum[r][c + 1] +
                        preSum[r + 1][c] -
                        preSum[r + 1][c + 1] +
                        (pizza[r].charAt(c) == 'A' ? 1 : 0);

        return dfs(m, n, k - 1, 0, 0, dp, preSum);
    }

    int dfs(int m, int n, int k, int r, int c, Integer[][][] dp, int[][] preSum) {
        if (preSum[r][c] == 0)
            return 0;

        if (k == 0)
            return 1;

        if (dp[k][r][c] != null)
            return dp[k][r][c];

        int ans = 0;

        for (int nr = r + 1; nr < m; nr++)
            if (preSum[r][c] - preSum[nr][c] > 0)
                ans = (ans + dfs(m, n, k - 1, nr, c, dp, preSum)) % 1_000_000_007;

        for (int nc = c + 1; nc < n; nc++)
            if (preSum[r][c] - preSum[r][nc] > 0)
                ans = (ans + dfs(m, n, k - 1, r, nc, dp, preSum)) % 1_000_000_007;

        return dp[k][r][c] = ans;
    }

}
